requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   google-genai
    #   httpx
    #   openai
cachetools==5.5.2
    # via google-auth
certifi==2025.4.26
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.2
    # via requests
deprecated==1.2.18
    # via
    #   opentelemetry-api
    #   opentelemetry-semantic-conventions
distro==1.9.0
    # via openai
ffmpeg-python==0.2.0
    # via poll-story-telegram-bot (pyproject.toml)
future==1.0.0
    # via ffmpeg-python
google-auth==2.40.2
    # via google-genai
google-genai==1.16.1
    # via poll-story-telegram-bot (pyproject.toml)
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   google-genai
    #   openai
    #   python-telegram-bot
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
importlib-metadata==8.6.1
    # via opentelemetry-api
jiter==0.10.0
    # via openai
openai==1.81.0
    # via poll-story-telegram-bot (pyproject.toml)
opentelemetry-api==1.33.1
    # via
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-httpx
    #   opentelemetry-semantic-conventions
opentelemetry-instrumentation==0.54b1
    # via opentelemetry-instrumentation-httpx
opentelemetry-instrumentation-httpx==0.54b1
    # via poll-story-telegram-bot (pyproject.toml)
opentelemetry-semantic-conventions==0.54b1
    # via
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-httpx
opentelemetry-util-http==0.54b1
    # via opentelemetry-instrumentation-httpx
packaging==25.0
    # via opentelemetry-instrumentation
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pydantic==2.11.4
    # via
    #   google-genai
    #   openai
pydantic-core==2.33.2
    # via pydantic
python-dotenv==1.1.0
    # via poll-story-telegram-bot (pyproject.toml)
python-telegram-bot==22.1
    # via poll-story-telegram-bot (pyproject.toml)
pyyaml==6.0.2
    # via poll-story-telegram-bot (pyproject.toml)
requests==2.32.3
    # via google-genai
rsa==4.9.1
    # via google-auth
sniffio==1.3.1
    # via
    #   anyio
    #   openai
tqdm==4.67.1
    # via openai
typing-extensions==4.13.2
    # via
    #   google-genai
    #   openai
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
urllib3==2.4.0
    # via requests
websockets==15.0.1
    # via google-genai
wrapt==1.17.2
    # via
    #   deprecated
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-httpx
zipp==3.21.0
    # via importlib-metadata
